Fix Edit Mode in Customization and Add User Role Modal Conditional

+++ components/feature/allPocket/DeleteLeavePocketAlert.jsx
@@ -19,18 +19,22 @@ export default function DeletePocketAlert({
   isOpen,
   onClose,
   onDelete,
+  onLeave,
   pocketName,
   pocketType,
   icon,
   color,
+  userRole = "owner",
 }) {
+  const isOwner = userRole === "owner";
+
   return (
     <AlertDialog isOpen={isOpen} onClose={onClose} size="md">
       <AlertDialogBackdrop />
       <AlertDialogContent className="bg-white px-6 py-8 rounded-2xl items-center justify-center">
         <AlertDialogHeader>
           <Heading size="xl" className="text-center font-bold">
-            Hapus pocket?
+            {isOwner ? "Hapus" : "Keluar dari"} pocket?
           </Heading>
         </AlertDialogHeader>
         <AlertDialogBody className="mb-6">
@@ -64,10 +68,14 @@ export default function DeletePocketAlert({
           {/* Delete */}
           <PrimaryButton
             buttonAction={() => {
-              onDelete();
+              if (isOwner) {
+                onDelete();
+              } else {
+                onLeave();
+              }
               onClose();
             }}
-            buttonTitle="Hapus pocket"
+            buttonTitle={`${isOwner ? "Hapus pocket" : "Keluar sekarang"}`}
             className="w-full bg-red-wondr mb-1 text-white active:bg-red-wondr-dark"
             textClassName="text-white"
           />
